+++ buildSrc/src/main/kotlin/Versions.kt
@@ -1,7 +1,7 @@
 object Versions {
     const val protobuf = "3.6.1"
-    const val grpc = "1.15.1"
-    const val kotlin = "1.3.11"
-    const val coroutines = "1.1.0"
+    const val grpc = "1.20.0"
+    const val kotlin = "1.3.31"
+    const val coroutines = "1.2.1"
     const val mockk = "1.9.1"
 }